+++ app/Http/Controllers/Controller.php
@@ -25,7 +25,7 @@ class Controller extends BaseController
         if (FacadesGate::allows('asAdmin')) {
             return redirect()->intended(Fortify::redirects('home', route('admin.beranda')));
         } else if (FacadesGate::allows('asUser')) {
-            return redirect()->intended(Fortify::redirects('home', 'home-user'));
+            return redirect()->intended(Fortify::redirects('home', route('user.index')));
         } else {
             return redirect('/');
         }
