+++ app/Http/Controllers/UserController.php
@@ -4,6 +4,8 @@ namespace App\Http\Controllers;
 
 use App\Models\Penyakit;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate as FacadesGate;
+
 
 class UserController extends Controller
 {
@@ -11,6 +13,12 @@ class UserController extends Controller
     {
         //get id, name, reason, solution, image from penyakit models
         $penyakit = Penyakit::get(['id', 'name', 'reason', 'solution', 'image']);
-        return view('layouts.user.app', compact('penyakit'));
+
+        return view('user.user', compact('penyakit'));
+    }
+
+    public function diagnosis()
+    {
+        return view('user.diagnosis');
     }
 }
