--- resources/views/user/beranda.blade.php
@@ -1,21 +1,0 @@
-<div id="beranda" class="row min-vh-100 align-content-center section">
-    <div class="col-12 col-md-6 py-5 " data-aos="fade-right" data-aos-anchor="body" id="col1">
-        <img class="img-fluid bg-body-tertiary rounded" id="gambar-cabai" src="{{ asset('assets/img/gambar-cabai.jpg') }}"
-            alt="Gambar Cabai https://fumida.co.id/wp-content/uploads/2021/03/67.-membasmi-hama-cabai.jpg">
-    </div>
-    <div class="col-12 col-md-6 align-self-center px-3 px-sm-5" data-aos="fade-left" data-aos-anchor="body"
-        id="col2">
-        <h1 class="text-start font-bold ">
-            {{-- <i class="fa-solid fa-pepper-hot fa-bounce fa-lg" style="color: #ff0000;"></i> --}}
-            Sistem Pakar
-            Diagnosa
-            Penyakit
-            Tanaman Cabai
-        </h1>
-        <p>
-            Lorem ipsum dolor sit amet consectetur adipisicing elit. Quas debitis quidem nostrum sed
-            id molestiae atque cumque, minima asperiores deserunt. Deserunt odit soluta nulla
-            praesentium quam necessitatibus incidunt ea laudantium!
-        </p>
-    </div>
-</div>
