--- resources/views/user/kontak.blade.php
@@ -1,101 +1,0 @@
-<div id="kontak" class="section" data-aos="fade-up">
-    <h2 class="fw-semibold pb-3">
-        Kontak
-    </h2>
-    <div class="row">
-        <div class="col-12 col-sm-7 pb-4" data-aos="fade-right">
-            <div class="card shadow">
-                <div class="card-body">
-                    <h4 class="font-semibold pb-2 card-title">
-                        Pengembang
-                    </h4>
-                    <div class="row">
-                        <div class="col-lg-6 col-12">
-                            <div class="image-slider">
-                                <div id="carouselExample" class="carousel slide">
-                                    <div class="carousel-inner">
-                                        <div class="carousel-item active" id="people1">
-                                            <img src="https://images.unsplash.com/photo-1438761681033-6461ffad8d80?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=870&q=80"
-                                                class="img-fluid" alt="...">
-                                        </div>
-                                        <div class="carousel-item" id="people2">
-                                            <img src="https://images.unsplash.com/photo-1534528741775-53994a69daeb?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=464&q=80"
-                                                class="img-fluid" alt="...">
-                                        </div>
-                                    </div>
-                                    <button class="carousel-control-prev button-slide" type="button"
-                                        data-bs-target="#carouselExample" data-bs-slide="prev">
-                                        <span class="carousel-control-prev-icon" aria-hidden="true"></span>
-                                        <span class="visually-hidden">Previous</span>
-                                    </button>
-                                    <button class="carousel-control-next button-slide" type="button"
-                                        data-bs-target="#carouselExample" data-bs-slide="next">
-                                        <span class="carousel-control-next-icon" aria-hidden="true"></span>
-                                        <span class="visually-hidden">Next</span>
-                                    </button>
-                                </div>
-                            </div>
-                        </div>
-                        <div class="col-lg-6 col-12 pt-3 pt-lg-0 align-self-center">
-                            <div class="row pb-2" id="descPeople">
-                                <div class="card-text" id="descPeople1">
-                                    <p class="mb-0 font-medium">
-                                        Rizky Ilham Pratama
-                                    </p>
-                                    <div class="col d-grid py-2">
-                                        <a class="btn btn-outline-dark text-nowrap text-start"
-                                            href="mailto:rizkyilhamp16@gmail.com" type="button">
-                                            <i class="fa-solid fa-envelope"></i>
-                                            rizkyilhamp16@gmail.com
-                                        </a>
-                                    </div>
-                                    <div class="col d-grid">
-                                        <a class="btn btn-outline-dark text-nowrap text-start" target="_blank"
-                                            type="button" href="https://instagram.com/rizkyilhampra">
-                                            <i class="fa-brands fa-instagram"></i>
-                                            @rizkyilhamp16
-                                        </a>
-                                    </div>
-                                </div>
-                                <div class="card-text d-none" id="descPeople2">
-                                    <p class="mb-0 font-medium">
-                                        Rizaldy Fauzan
-                                    </p>
-                                    <div class="col py-2 d-grid">
-                                        <a class="btn btn-outline-dark text-nowrap text-start" href="#">
-                                            <i class="fa-solid fa-envelope"></i>
-                                            rizaldyfauzan@gmail.com
-                                        </a>
-                                    </div>
-                                    <div class="col d-grid">
-                                        <a class="btn btn-outline-dark text-nowrap text-start" target="_blank"
-                                            href="#">
-                                            <i class="fa-brands fa-instagram"></i>
-                                            @rizaldyfauzan
-                                        </a>
-                                    </div>
-                                </div>
-                            </div>
-                        </div>
-                    </div>
-                </div>
-
-            </div>
-
-        </div>
-        <div class="col-12 col-sm-5" data-aos="fade-left">
-            <div class="card shadow">
-                <div class="card-body">
-                    <h4 class="font-semibold pb-2 card-title">
-                        Tentang Kami
-                    </h4>
-                    <div class="card-text">
-                        Lorem ipsum, dolor sit amet consectetur adipisicing elit. Molestiae maiores nisi provident rerum
-                        officiis, perspiciatis quae totam pariatur sint rem ullam impedit omnis vitae recusandae, ipsum
-                        dicta deserunt sunt cum.
-                    </div>
-                </div>
-            </div>
-        </div>
-    </div>
-</div>
